Add a maxLayers setting to the GIS configuration.
It is useful to limit the maximum numer of splitting layers
in GIS. It used to be hard-coded to 10 when running in the
speed-distribution-only mode, and to 25 when running in the full limit
calculation mode. This commit makes it configurable by introducing the
maxLayers int setting into the config. If maxLayers is not provided in
the config file, it is set to 25.